Move the yaml loading into the model
The model represents the config file entry and thus is the logical place
to do the YAML load.

NOTE: The Batch model is only added to the session (and thus saved) if
it is actually ran
